list/serializers/user_serializer.py
from rest_framework import serializers
from list.models import CustomUser


class CustomUserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = '__all__'
