requirements.txt
django-cors-headers==4.3.1
celery==5.4.0
django-celery-beat==2.6.0
django==5.0.4
APScheduler==3.10.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
psycopg2==2.9.9
django-crispy-forms==1.14.0
whitenoise==6.7.0
#crispy-bootstrap4==2024.1
